RootOutput: Small changes
* change parameter type to unsigned and large enough to contain any
  useful size
* warn if split size >= 1 TB is given
* expand parameter documentation
* use braces where possible
* move dereference to variable ... by using && instead of and
* adjust log messages

  * info for created directories
  * debug for writing trees
  * update debug levels

+++ framework/modules/rootio/include/RootOutputModule.h
@@ -156,7 +156,7 @@ namespace Belle2 {
 
     /** Maximum output file size in MB. If not set we don't split. Otherwise we split
      * if the event tree in output file has reached the given size in MB */
-    boost::optional<int> m_outputSplitSize{boost::none};
+    boost::optional<uint64_t> m_outputSplitSize{boost::none};
 
     //then those for purely internal use:
